Rename non-generated `fix` package to `metafix` (#35)
