--- org.metafacture.fix.ide/src/main/java/org/metafacture/fix/ide/FixIdeModule.java
@@ -1,11 +1,0 @@
-package org.metafacture.fix.ide;
-
-/**
- * Use this class to register ide components.
- */
-public class FixIdeModule extends AbstractFixIdeModule {
-
-    public FixIdeModule() {
-    }
-
-}
