+++ org.metafacture.fix.ide/src/main/java/org/metafacture/metafix/ide/FixIdeModule.java
@@ -1,0 +1,9 @@
+package org.metafacture.metafix.ide;
+
+/**
+ * Use this class to register ide components.
+ */
+public class FixIdeModule extends AbstractFixIdeModule {
+    public FixIdeModule() {
+    }
+}
