+++ org.metafacture.fix.web/build.gradle
@@ -27,7 +27,7 @@ dependencies {
 task jettyRun(type: JavaExec) {
   dependsOn(sourceSets.main.runtimeClasspath)
   classpath = sourceSets.main.runtimeClasspath.filter { it.exists() }
-  main = 'org.metafacture.fix.web.ServerLauncher'
+  main = 'org.metafacture.metafix.web.ServerLauncher'
   standardInput = System.in
   group = 'run'
   description = 'Starts an example Jetty server with your language'
