--- org.metafacture.fix.web/src/main/java/org/metafacture/fix/web/FixWebModule.java
@@ -1,11 +1,0 @@
-package org.metafacture.fix.web;
-
-/**
- * Use this class to register additional components to be used within the web application.
- */
-public class FixWebModule extends AbstractFixWebModule {
-
-    public FixWebModule() {
-    }
-
-}
