+++ org.metafacture.fix.web/src/main/java/org/metafacture/metafix/web/FixWebModule.java
@@ -1,0 +1,10 @@
+package org.metafacture.metafix.web;
+
+/**
+ * Use this class to register additional components to be used within the web
+ * application.
+ */
+public class FixWebModule extends AbstractFixWebModule {
+    public FixWebModule() {
+    }
+}
