--- org.metafacture.fix/src/main/java/org/metafacture/fix/validation/FixValidator.java
@@ -1,13 +1,0 @@
-package org.metafacture.fix.validation;
-
-/**
- * This class contains custom validation rules.
- *
- * See https://www.eclipse.org/Xtext/documentation/303_runtime_concepts.html#validation
- */
-public class FixValidator extends AbstractFixValidator {
-
-    public FixValidator() {
-    }
-
-}
