+++ org.metafacture.fix/src/main/java/org/metafacture/metafix/GenerateFix.mwe2
@@ -1,0 +1,51 @@
+module org.metafacture.metafix.GenerateFix
+
+import org.eclipse.xtext.xtext.generator.*
+import org.eclipse.xtext.xtext.generator.model.project.*
+
+Workflow {
+  component = XtextGenerator {
+    configuration = {
+      project = StandardProjectConfig {
+        baseName = "org.metafacture.fix"
+        rootPath = ".."
+        runtimeTest = {
+          enabled = true
+        }
+        web = {
+          enabled = true
+        }
+        mavenLayout = true
+      }
+      code = {
+        encoding = "UTF-8"
+        lineDelimiter = "\n"
+        fileHeader = "/*\n * generated by Xtext \${version}\n */"
+        preferXtendStubs = false
+      }
+    }
+    language = StandardLanguage {
+      name = "org.metafacture.metafix.Fix"
+      fileExtensions = "fix"
+      serializer = {
+        generateStub = false
+      }
+      validator = {
+        // Generates checks for @Deprecated grammar annotations, an IssueProvider and a corresponding PropertyPage
+        generateDeprecationValidation = true
+      }
+      junitSupport = {
+        junitVersion = "5"
+        generateStub = false
+      }
+      webSupport = {
+        framework = "ace"
+        generateServlet = false
+        generateJettyLauncher = false
+      }
+      xbaseSupport = {
+        generateXtendInferrer = false
+      }
+    }
+  }
+}
