+++ org.metafacture.fix/src/main/java/org/metafacture/metafix/scoping/FixScopeProvider.java
@@ -1,0 +1,14 @@
+package org.metafacture.metafix.scoping;
+
+/**
+ * This class contains custom scoping description.
+ *
+ * See https://www.eclipse.org/Xtext/documentation/303_runtime_concepts.html#scoping
+ * on how and when to use it.
+ */
+public class FixScopeProvider extends AbstractFixScopeProvider {
+
+    public FixScopeProvider() {
+    }
+
+}
