+++ org.metafacture.fix/src/main/java/org/metafacture/metafix/validation/FixValidator.java
@@ -1,0 +1,13 @@
+package org.metafacture.metafix.validation;
+
+/**
+ * This class contains custom validation rules.
+ *
+ * See https://www.eclipse.org/Xtext/documentation/303_runtime_concepts.html#validation
+ */
+public class FixValidator extends AbstractFixValidator {
+
+    public FixValidator() {
+    }
+
+}
